Performs Mechanics Checks on a MDS Lab This function check that you have a Github repo link, that you have pushed your latest commit, and that you have at least three commit messages authored by you in your history.

Performs Mechanics Checks on a MDS Lab This function check that you have a Github repo link, that you have pushed your latest commit, and that you have at least three commit messages authored by you in your history.

check_mechanics(
  notebook = NULL,
  repo = ".",
  branch = usethis::git_branch_default()
)

Arguments

notebook
A character indicating the path or list of paths to MDS lab files (either.ipynb or .Rmd). If left blank, the function will recursively search for all labs in the working directory based on the file extension.

repo
A character path to the repo to check. Defaults to the working directory.

branch
A string specifying the name of the default branch used for grading.

Value

The function prints the results of the mechanics checks to screen. Silently returns TRUE if all the checks are passed.

Examples

if (FALSE) {
check_mechanics()
}