Parse MDS lab files to return the markdown content

Parse MDS lab files to return the markdown content

parse_lab(notebook = NULL)

Arguments

notebook
A character indicating the path or list of paths to MDS lab files (either.ipynb or .Rmd). If left blank, the function will recursively search for all labs in the working directory based on the file extension.

Value

A character vector where each element is the content of one markdown cell.

Examples

# R markdown
notebook <- system.file("extdata", "dummylab.Rmd", package = "labzenr")
parse_lab(notebook)

#> [1] "---\ntitle: DSCI 999 Lab 2 - Another Lab\noutput:\n  html_document:\n    toc: true\n    toc_depth: 2\n---\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [2] "\n\nThis lab has a total of 20 points: 14 required, and 6 optionals.\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Donec at massa a quam posuere elementum at nec lectus. Vivamus in pharetra neque. Vivamus sit amet enim sed lorem imperdiet pretium. Vestibulum dignissim facilisis neque non rhoncus. Vestibulum facilisis egestas nulla in molestie. Pellentesque vel tempus magna. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque vitae nibh sollicitudin, rhoncus nunc at, sollicitudin nibh. Fusce at enim consectetur tortor aliquam maximus ac ullamcorper leo. Proin consectetur nulla et rhoncus posuere. Proin gravida ut mi quis elementum. Suspendisse laoreet, eros in consectetur tristique, massa velit maximus sapien, sit amet porta dui ex sed leo. Etiam id est arcu.\n\n\n## General Instructions\n\nrubric={mechanics:3}\n\n> The Github repo for this lab can be found at [MDS-2021-22/DSCI_999_lab2_johnsmith](https://github.ubc.ca/MDS-2020-21/DSCI_999_lab2_johnsmith).\n\n-   All files necessary to run your work must be pushed to your **GitHub.ubc.ca** repository for this lab.\n-   You need to have a minimum of 3 commit messages associated with your **GitHub.ubc.ca** repository for this lab.\n-   You must also submit a `.html` version of this homework to Canvas.\n-   Follow the [MDS general lab instructions](https://ubc-mds.github.io/resources_pages/general_lab_instructions/).\n-   This assignment is to be completed in `R`.\n-   Make it easy for others to read and run your code. Use the *pipe* operator `%>%` whenever appropriate to help with code readability.\n-   For coding questions, ensure you display the output of the code object asked for in the questions.\n\n## Exercise 1: True or False?\n<hr>\nrubric={reasoning:5}\n\nCurabitur in mauris id sapien accumsan pretium nec non leo. Cras ipsum nulla, condimentum eget vulputate eu, rutrum vel nisl. Pellentesque luctus velit eu lacus malesuada auctor. Sed lobortis mauris eget sem fermentum lobortis. Cras bibendum magna risus, tincidunt placerat ex hendrerit et. Aenean sit amet lorem ac ante egestas pretium. Donec sed nisi risus. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Nulla ut erat id massa imperdiet congue sed ut lacus.\n\n\n\n> Your answer goes here.\n\n\n\n## Exercise 2: Exploratory Analysis\n\n### 2(a)\nrubric={reasoning:6}\n\n"
#> [3] "\n\n\n\n### Optional 2(b)\n rubric={reasoning:3}\n\nThis one has a space before the rubric, and some latex blocks:\n\n$$\ny_i \\sim \\textrm{Normal}(0, 1)\n$$\n\n> Your answer goes here.\n\n\n\n### Optional 2(c)\n<br>\nrubric={reasoning:3}\n\nQuisque feugiat varius nunc ac pellentesque. Mauris pretium ante id nibh malesuada aliquam. Suspendisse bibendum ex sed sollicitudin bibendum. Suspendisse potenti. Donec auctor pharetra ultricies. Aliquam erat volutpat. Praesent in tincidunt nisl, sit amet blandit augue. Phasellus et est mollis, imperdiet orci ac, scelerisque est. Nullam mattis ultricies porttitor. Donec eu euismod purus, vitae luctus ipsum. Maecenas vitae ullamcorper felis. Integer felis sapien, commodo at metus ut, hendrerit imperdiet elit. In quis posuere libero.\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [4] "\n\n\n> Your answer goes here.\n\n\n# Submission\n\nCongratulations! You are done the lab! Pat yourself on the back, render your notebook to `html` by clicking the knit button inside RStudio and submit your worksheet to both Canvas and GitHub!\n\nTo programmatically submit the assignment to Canvas, open [submit.ipynb](submit.ipynb) and use the instructions there."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           

# Python notebook
notebook <- system.file("extdata", "dummylab.ipynb", package = "labzenr")
parse_lab(notebook)

#>  [1] "# DSCI 999 - DUMMY LAB\n\n\n\n# Lab 1: A dummy lab to test the parsing features\n\n\n\nIf this lab is properly parsed, there should be 19 required points, 20 optional points, and 39 points total."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [2] "## Table of contents\n\n- [Submission instructions](#si) (5%)\n\n- [Imports](#im)\n\n- [Exercise 1: Warm-up](#1) (35%)"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [3] "## Submission instructions <a name=\"si\"></a>\n\n<hr>\n\nrubric={mechanics:5}\n\n\n\nIn this example, the rubric follows \n\n\n\nYou will receive marks for correctly submitting this assignment. \n\n\n\nTo correctly submit this assignment follow the instructions below:\n\n\n\n- Push your assignment to your GitHub repository. \n\n- Add a link to your GitHub repository here: **INSERT LINK HERE**\n\n- Upload an HTML render of your assignment to Canvas. The last cell of this notebook will help you do that.\n\n- Be sure to follow the [general lab instructions](https://ubc-mds.github.io/resources_pages/general_lab_instructions/).\n\n\n\n[Here](https://github.com/UBC-MDS/public/tree/master/rubric) you will find the description of each rubric used in MDS.\n\n\n\n**NOTE: The data you download for use in this lab SHOULD NOT BE PUSHED TO YOUR REPOSITORY. You might be penalised for pushing datasets to your repository. I have seeded the repository with `.gitignore` and hoping that it won't let you push CSVs.**"                                                                                                                       
#>  [4] "<br><br><br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#>  [5] "## Imports <a name=\"im\"></a>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  [6] "<br><br><br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#>  [7] "## Exercise 1: Warm-up <a name=\"1\"></a>\n\n<hr>\n\n\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Donec at massa a quam posuere elementum at nec lectus. Vivamus in pharetra neque. Vivamus sit amet enim sed lorem imperdiet pretium. Vestibulum dignissim facilisis neque non rhoncus. Vestibulum facilisis egestas nulla in molestie. Pellentesque vel tempus magna. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque vitae nibh sollicitudin, rhoncus nunc at, sollicitudin nibh. Fusce at enim consectetur tortor aliquam maximus ac ullamcorper leo. Proin consectetur nulla et rhoncus posuere. Proin gravida ut mi quis elementum. Suspendisse laoreet, eros in consectetur tristique, massa velit maximus sapien, sit amet porta dui ex sed leo. Etiam id est arcu."                                                                                                                                                                                                                                                                                                                                                          
#>  [8] "### 1.1 Data Loading and Exploration\n\nrubric={reasoning:4}\n\n\n\nNote, the above code is the simplest, as the rubric follows right below the heading. \n\n\n\n**Your tasks:**\n\n\n\nRun the code below and answer the following question. \n\n\n\n1. What is your favourite animal?   \n\n\n\n> *Note 1: Some comment\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [9] "### Solution_1_1_1\n\n\n\n\n\nYour answer here."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [10] "<br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [11] "### 1.2 A second exercise\n\n\n\nrubric={accuracy:3,reasoning:2}\n\n\n\nIn this example, there is an extra line between the header and the rubric. The rubric has two components."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> [12] "### Solution_1_2_2\n\n\n\n\n\nYour answer here."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [13] "<br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [14] "### 1.3 (optional) A third exercise\n\nrubric={accuracy:6,reasoning:4}\n\n\n\nCurabitur in mauris id sapien accumsan pretium nec non leo. Cras ipsum nulla, condimentum eget vulputate eu, rutrum vel nisl. Pellentesque luctus velit eu lacus malesuada auctor. Sed lobortis mauris eget sem fermentum lobortis. Cras bibendum magna risus, tincidunt placerat ex hendrerit et. Aenean sit amet lorem ac ante egestas pretium. Donec sed nisi risus. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Nulla ut erat id massa imperdiet congue sed ut lacus.\n\n\n\nQuisque feugiat varius nunc ac pellentesque. Mauris pretium ante id nibh malesuada aliquam. Suspendisse bibendum ex sed sollicitudin bibendum. Suspendisse potenti. Donec auctor pharetra ultricies. Aliquam erat volutpat. Praesent in tincidunt nisl, sit amet blandit augue. Phasellus et est mollis, imperdiet orci ac, scelerisque est. Nullam mattis ultricies porttitor. Donec eu euismod purus, vitae luctus ipsum. Maecenas vitae ullamcorper felis. Integer felis sapien, commodo at metus ut, hendrerit imperdiet elit. In quis posuere libero."
#> [15] "### Solution_1_3_1\n\n\n\nYour answer here."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [16] "<br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [17] "### 1.4 Another Optional Exercise\n\n<br>\n\nrubric={accuracy:7,reasoning:3}\n\n\n\n\n\nNote how this one is capitalized, and that there is a `<br>` tag between the rubric and header"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [18] "<br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [19] "### 1.5 Visualizing results \n\n rubric={viz:5}\n\n\n\nIn this example, there is a leading space in front of \"rubric\"\n\n\n\nIn eget elit interdum, tincidunt augue in, fringilla eros. Aenean congue porttitor turpis nec tempus. Suspendisse blandit, velit sed vehicula tincidunt, enim dui luctus orci, nec volutpat diam augue vel neque. Nam scelerisque luctus risus, sed ornare turpis euismod vel. Duis ultricies lectus non ipsum imperdiet, quis dictum elit vulputate. Aenean quis arcu hendrerit, maximus justo vel, commodo lorem. Vivamus in lacinia turpis, a accumsan purus. Sed ultricies nunc sit amet pulvinar fermentum. Vivamus tempus, magna vitae varius sagittis, leo metus convallis eros, non condimentum ligula ligula at mauris. Vivamus fermentum nibh quis tortor porttitor vulputate. Vestibulum lacus nisl, condimentum ac sagittis mattis, laoreet in risus. Aliquam erat volutpat. Vestibulum finibus purus sed enim fringilla, et cursus leo ornare. "                                                                                                                                                                                
#> [20] "<br><br><br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [21] "<br><br>"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [22] "### A side note \n\nrubric is a word that can be a trap, if it is placed at the start of a sentence but do not actually convey points.\n\n\n\nThey can also be used in the middle of a sentence, like when I mention a rubric here."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [23] "### Submission to Canvas\n\n\n\n**PLEASE READ: When you are ready to submit your assignment do the following:**\n\n\n\n- Run all cells in your notebook to make sure there are no errors by doing Kernel -->  Restart Kernel and Run All Cells...\n\n- Make sure that all plots are visible in your HTML render. \n\n- If you are using the \"999\" `conda` environment, make sure to select it before running all cells. \n\n- Convert your notebook to .html format using the `convert_notebook()` function below or by File -> Export Notebook As... -> Export Notebook to HTML\n\n- Run the code `submit()` below to go through an interactive submission process to Canvas.\n\nAfter submission, be sure to do a final push of all your work to GitHub (including the rendered html file)."                                                                                                                                                                                                                                                                                                                                                                            
#> [24] "Well done!! Congratulations on finishing the lab!! "